Initial commit of JavaScript Tile Swap game
Two by two version of the puzzle that is playable: you can only swap tiles according to the rules, it tracks the number of moves, and reports when you solved the puzzle. The visualisation is still basic. Only a simple table, with no effort spend on making it look good. It will be replaced anyway.